requirements.txt
Django==4.0.3
djangorestframework==3.13.1
PyMySQL==1.0.2
python-dotenv==0.19.2
django-cors-headers==3.11.0
django-heroku==0.3.1
